feat: add redux api query
