+++ src/api/api.ts
@@ -1,29 +1,12 @@
-import { CardPreviewData, CardFullData } from 'types';
+import { createApi, fetchBaseQuery } from '@reduxjs/toolkit/query/react';
+import { CardFullData } from 'types';
 
-const BASE_URL = 'https://rickandmortyapi.com/api';
-
-export const getCharacters = async (
-  searchValue?: string
-): Promise<CardPreviewData[] | undefined> => {
-  try {
-    let name;
-    if (searchValue) {
-      name = `?name=${searchValue}`;
-    }
-    const res = await fetch(`${BASE_URL}/character${name || ''}`);
-    const data = await res.json();
-    return data.results;
-  } catch (error) {
-    console.log(error);
-  }
-};
-
-export const getSingleCharacter = async (id: number): Promise<CardFullData | undefined> => {
-  try {
-    const res = await fetch(`${BASE_URL}/character/${id}`);
-    const data = await res.json();
-    return data;
-  } catch (error) {
-    console.log(error);
-  }
-};
+export const rickAndMortyApi = createApi({
+  reducerPath: 'rickAndMortyApi',
+  baseQuery: fetchBaseQuery({ baseUrl: 'https://rickandmortyapi.com/api/' }),
+  endpoints: (builder) => ({
+    getSingleCharacter: builder.query<CardFullData, string>({
+      query: (id) => `character/${id}`,
+    }),
+  }),
+});
