+++ src/pages/HomePage/HomePage.test.tsx
@@ -1,7 +1,7 @@
 import { test, expect, vi, Mock } from 'vitest';
 import { render, screen, act, cleanup } from '@testing-library/react';
 import { CardPreviewData } from 'types';
-import { getCharacters } from '../../api/api';
+import { getCharacters } from '../../api/apiOld';
 
 import React from 'react';
 import HomePage from './HomePage';
