+++ src/store/index.ts
@@ -1,0 +1,8 @@
+import { configureStore } from '@reduxjs/toolkit';
+import { rickAndMortyApi } from 'api/api';
+
+export const store = configureStore({
+  reducer: {
+    [rickAndMortyApi.reducerPath]: rickAndMortyApi.reducer,
+  },
+});
